Add settings to control adaptive batching
