+++ mlserver/batching/__init__.py
@@ -1,4 +1,4 @@
 from .adaptive import AdaptiveBatcher
-from .listeners import load_batching
+from .hooks import load_batching
 
 __all__ = ["AdaptiveBatcher", "load_batching"]
