+++ tests/testdata/model-settings.json
@@ -3,6 +3,9 @@
 
   "implementation": "tests.fixtures.SumModel",
 
+  "max_batch_size": 10,
+  "max_batch_time": 0.4,
+
   "versions": ["sum-model/v1.2.3"],
   "platform": "mlserver",
   "inputs": [
